Feat #43 : 회원가입 이메일 인증 Toast 추가 및 리팩토링
