--- src/Pages/SignUp/Components/SignUpForm/types.ts
@@ -1,3 +1,0 @@
-export interface SignUpFormProps {
-  select: boolean;
-}
