+++ src/Pages/SignUp/index.tsx
@@ -1,24 +1,4 @@
-import { createContext, useState } from 'react';
-import InfoForm from './Components/InfoForm';
-import { SignUpWrapper } from './style';
+import InfoPage from './InfoPage';
+import SignUpPage from './SignUpPage';
 
-export const SignUpContext = createContext({
-  choose: '',
-  // eslint-disable-next-line no-unused-vars, @typescript-eslint/no-unused-vars
-  setChooseHandler: (event: string) => {},
-});
-
-const SignUpPage = () => {
-  const [choice, setChoice] = useState('');
-  const setChoiceHandler = (event: string) => setChoice(event);
-
-  return (
-    <SignUpWrapper>
-      <SignUpContext.Provider value={{ choose: choice, setChooseHandler: setChoiceHandler }}>
-        <InfoForm />
-      </SignUpContext.Provider>
-    </SignUpWrapper>
-  );
-};
-
-export default SignUpPage;
+export { InfoPage, SignUpPage };
